pyproject.toml
[build-system]
requires = ["setuptools~=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "regula-ai"
version = "1.2.0"
description = "AI governance risk indication for code — EU AI Act compliance at the point of creation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "The Implementation Layer"}]
keywords = ["ai-governance", "eu-ai-act", "compliance", "claude-code", "risk-indication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
]

[project.optional-dependencies]
yaml = ["pyyaml~=6.0"]
ast = ["tree-sitter~=0.23", "tree-sitter-javascript~=0.23", "tree-sitter-typescript~=0.23"]
test = ["pytest>=7.0"]
pdf = ["weasyprint~=62.0"]
all = ["pyyaml~=6.0", "tree-sitter~=0.23", "tree-sitter-javascript~=0.23", "tree-sitter-typescript~=0.23", "weasyprint~=62.0"]

[project.urls]
Homepage = "https://github.com/kuzivaai/getregula"
Repository = "https://github.com/kuzivaai/getregula"
Issues = "https://github.com/kuzivaai/getregula/issues"

[project.scripts]
regula = "scripts.cli:main"

[tool.setuptools.packages.find]
include = ["scripts*", "hooks*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
